inmuebles_app/requirements.txt
Flask==3.0.3
mysql-connector-python==8.3.0
requests==2.32.3
pytest==8.3.5
pytest-cov==5.0.0